AI project
One of the goals of setting this assignment is to give you a program that you can be
proud of which you can use in job interviews and put in a portfolio.
What you will be building:
1. A system that takes a description of a television show and reports the genres of the
show.
2. A system that takes a description of a television show and finds the most likely
television show in its database.
For both tasks it is easy to get something simple working – this will be enough for a
pass. But it is also very hard to create something excellent that handles everything —
which you would need to do to get top marks.
The sample data for both sections will be a SQLite database, which is available on
ilearn. It is not the same data that your programs will be evaluated on, but the schema
will be the same.